StringDisplay is complete, extracted into its own file  and imported
