+++ src/Components/string-display/string-display.css
@@ -1,0 +1,16 @@
+/* ReverseString Formatting */
+.display {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 200px;
+  height: 100px;
+  align-items: center;
+  justify-content: center;
+}
