+++ src/index.css
@@ -13,19 +13,3 @@ code {
 }
 
 
-/* ReverseString Formatting */
-.display {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.container {
-  display: flex;
-  flex-direction: column;
-  width: 200px;
-  height: 100px;
-  align-items: center;
-  justify-content: center;
-}
